py-test-srv/pytest/bin/requirements.txt
# deps for app
bottle
bottle-sqlalchemy
pymysql

# deps for pytest
pytest
allure-pytest
